requirements.txt
psycopg2-binary==2.9.5
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.1
SQLAlchemy==1.4.46
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
